+++ src/main/java/module-info.java
@@ -22,10 +22,12 @@ module awt.extensions.main
 {
 	requires java.desktop;
 	requires lombok;
+	requires java.prefs;
 
 	exports io.github.astrapi69.awt.action;
 	exports io.github.astrapi69.awt.extension;
 	exports io.github.astrapi69.awt.screen;
 	exports io.github.astrapi69.awt.system;
 	exports io.github.astrapi69.awt.window.adapter;
+	exports io.github.astrapi69.awt.screen.position;
 }